# specs/001-ai-client-documentation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Warba Bank — AI-Powered Client Documentation API
  version: 1.0.0
  description: |
    Backend contract for Track 1 of the Warba Bank Corporate Banking AI Challenge.

    Governed by Constitution v1.0.0. Three contract-level invariants are worth stating
    before the endpoints, because they are the reason several endpoints look the way
    they do:

    1. **No auto-approval path exists.** `POST /documents/{id}/approve` is the only
       transition into `APPROVED`, it requires an authenticated RM, and it requires the
       caller to echo the `content_hash` of the exact version being approved. There is no
       endpoint, parameter, or scheduled process that approves a document any other way
       (Principle III, FR-026/FR-027).
    2. **Failure produces no document.** Generation, screening, and retrieval failures
       return a 4xx/5xx error, never a partial draft (Principle I, FR-037).
    3. **Audit is read-only over HTTP.** No endpoint mutates or deletes an audit event.
       The database role backing this API does not hold UPDATE or DELETE on the audit
       table (Principle VIII, FR-032).

    All data served by this API is synthetic (Principle VII).
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /documents/{document_id}/sections/{section_key}:
    patch:
      tags: [Sections]
      summary: Edit a section inline
      description: |
        Creates a new `RM_EDITED` version. The edit is attributed to the acting RM and is
        distinguishable from AI-generated content in the audit record (FR-022).

        Requires `If-Match` carrying the current version's `content_hash`. A stale value
        returns 412 rather than silently overwriting a concurrent edit (FR-040).
      parameters:
        - { name: document_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: section_key, in: path, required: true, schema: { type: string } }
        - name: If-Match
          in: header
          required: true
          schema: { type: string }
          description: content_hash of the version being edited
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content: { type: string }
                resolved_gaps:
                  type: array
                  items:
                    type: object
                    required: [field, resolution_note]
                    properties:
                      field: { type: string }
                      resolution_note: { type: string }
      responses:
        '200':
          description: Edited; new version created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentDetail' }
        '409':
          description: Document is APPROVED and cannot be edited
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '412':
          description: Stale content_hash — another session edited this document (FR-040)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '451':
          description: The edited content failed Shariah screening
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScreeningBlockedError' }
  /documents/{document_id}/sections/{section_key}/regenerate:
    post:
      tags: [Sections]
      summary: Regenerate a single section
      description: |
        Regenerates one section against the existing Evidence Ledger without discarding
        accepted content elsewhere in the document (FR-023). Creates a
        `REGENERATED_SECTION` version.
      parameters:
        - { name: document_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: section_key, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                instruction: { type: string, maxLength: 1000 }
      responses:
        '200':
          description: Section regenerated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentDetail' }
        '422':
          description: Regenerated section failed deterministic validation; prior version retained
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '451':
          description: Regenerated section failed Shariah screening
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScreeningBlockedError' }

  # ─────────────────────────── Approval ───────────────────────────
...
  /documents/{document_id}/export:
    get:
      tags: [Export]
      summary: Export an approved document
      description: |
        Only `APPROVED` documents export. The rendered file embeds the approval record
        (approver, timestamp, content hash), the Shariah review status, and AI-assisted
        attribution (FR-036). Exporting is itself an audit event.
      parameters:
        - { name: document_id, in: path, required: true, schema: { type: string, format: uuid } }
        - name: format
          in: query
          schema: { type: string, enum: [docx, pdf], default: docx }
      responses:
        '200':
          description: Rendered document
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema: { type: string, format: binary }
            application/pdf:
              schema: { type: string, format: binary }
        '409':
          description: Document is not APPROVED
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ───────────────────────────── Audit ────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    DocumentType:
      type: string
      enum: [CALL_REPORT, CLIENT_PROFILE, CREDIT_MEMO_NARRATIVE, KYC_SUMMARY]
    DocumentStatus:
      type: string
      enum: [DRAFT, UNDER_REVIEW, REJECTED, APPROVED]
    ShariahStatus:
      type: string
      enum: [PENDING_REVIEW, CLEARED, FLAGGED]
      description: |
        Defaults to PENDING_REVIEW and is never set to CLEARED by the system. The system
        prepares documents for Shariah review; it does not clear them (Principle II).
    Confidence:
      type: string
      enum: [HIGH, MEDIUM, LOW]
...
    GapMarker:
      type: object
      description: |
        A first-class output state, not an error. Rendered visibly to the RM; blocks approval
        until resolved or explicitly acknowledged (FR-025).
      required: [field, label, resolved]
      properties:
        field: { type: string }
        label:
          type: string
          example: '[MISSING: audited turnover FY2025]'
        resolved: { type: boolean }
        resolution_note: { type: string, nullable: true }
...
    DocumentSection:
      type: object
      required: [section_key, title, ordinal, evidence_refs, gaps, confidence]
      properties:
        section_key: { type: string }
        title: { type: string }
        ordinal: { type: integer }
        content:
          type: string
          nullable: true
          description: Null when the section is entirely a gap.
        evidence_refs:
          type: array
          items: { type: string }
          description: |
            claim_ids. Every one is validated against the ledger before the section is
            returned; an unresolvable reference converts the section to a gap (FR-011, R3).
        gaps:
          type: array
          items: { $ref: '#/components/schemas/GapMarker' }
        confidence: { $ref: '#/components/schemas/Confidence' }
        is_rm_edited: { type: boolean }
        contains_external_data: { type: boolean }
    DocumentSummary:
      type: object
      required: [id, client_id, document_type, status, shariah_status, created_at]
      properties:
        id: { type: string, format: uuid }
        client_id: { type: string, format: uuid }
        client_reference: { type: string }
        document_type: { $ref: '#/components/schemas/DocumentType' }
        status: { $ref: '#/components/schemas/DocumentStatus' }
        shariah_status: { $ref: '#/components/schemas/ShariahStatus' }
        created_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
    DocumentDetail:
      allOf:
        - $ref: '#/components/schemas/DocumentSummary'
        - type: object
          required: [version_number, content_hash, sections, ai_generated, unresolved_gap_count]
          properties:
            version_number: { type: integer }
            content_hash:
              type: string
              description: Echo this to PATCH (If-Match) and to approve.
            ai_generated:
              type: boolean
              enum: [true]
              description: AI-generated content is labelled at all times (FR-020).
            template_version: { type: string }
            prompt_version: { type: string }
            model_id: { type: string, nullable: true }
            sections:
              type: array
              items: { $ref: '#/components/schemas/DocumentSection' }
            unresolved_gap_count:
              type: integer
              description: Approval is blocked while this is greater than zero (FR-025).
            screening:
              type: object
              properties:
                outcome: { type: string, enum: [PASS, BLOCKED, FLAGGED] }
                vocabulary_version: { type: string }
                findings:
                  type: array
                  items:
                    type: object
                    properties:
                      term: { type: string }
                      section_key: { type: string }
                      severity: { type: string }
                      rule_id: { type: string }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message:
          type: string
          description: |
            Plain, non-technical, and actionable — the RM must understand what happened and
            what to do next (FR-038, NFR-UX-02).
        detail: { type: object, additionalProperties: true }
    ScreeningBlockedError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            vocabulary_version: { type: string }
            findings:
              type: array
              items:
                type: object
                properties:
                  term: { type: string }
                  section_key: { type: string }
                  severity: { type: string }
                  rule_id: { type: string }
                  rationale: { type: string }
...